src/pr-bot/requirements.txt
langchain-core 
langchain-community 
langchain-openai
PyGithub
GitPython
loguru